db: relax version checking
When calling `search_generic()`, the `version` parameter is taken from
the output of `cargo metadata`, which always return the latest version
available on crates.io. Due to ba11751 this implies the crate will be
considered missing from Debian unless the packaged version is the latest
one.

This commit adds a fallback version check in case the initial check
fails, by trimming the version to a supposedly compatible semver
version.

Fixes #31