db: add test for new version check logic
This new test searches for `serde` in Debian Bullseye, which is
`oldstable` now and thus won't be updated anymore. It requests various
versions to ensure the crate is reported as "found", "outdated" or
"compatible" appropriately. It also searches for the non-existing
`notacrate` to ensure it properly reports missing packages.

Finally, this test (which requires network access) is marked as ignored
so it isn't run by default, avoiding issues with Debian's build process.
The workflow has been updated to ensure this test is run in CI.