test: adapt to workspace series
This commit adapts all pre-existing tests to the workspace series mainly by
replacing the 🔴 icon with 🪏  for nodes which are part of the current
workspace. In print_tree_with_common_dependency and
print_dependency_status_in_debian, we now need to pass a `-w` flag, otherwise
the tree of each workspace member is printed. This is the new intended behavior
(but does not actually change the old behavior as workspaces were not previously
supported anyway).