fix: Remove `--color` CLI option
This option is only passed to `cargo`, but it is useless to do so
since the dependency is rendered by `cargo-debstatus`, not `cargo`.
Colors seem to get deactivated when the standard output is piped into
a file, so there does not seem to be a need to provide this switch.