fix: also search for packages via 'Provides:' field
For #71.

This is unfortunately very slow as Provides are only stored in UDD
in a big text field. This means that running cargo debstatus on a
project with many missing dependencies will run quite a few expensive
queries. This should either be opt-in, or use a more efficient way of
querying UDD.